Fix imports, fix tests
