+++ api/server/routes/aasx_file_server.py
@@ -3,7 +3,7 @@ from typing import Any
 from aas_core3.types import Identifiable
 from fastapi import APIRouter, Request
 
-from server.services.aasx_flie_server_service import AasxFileServerService
+from server.services.aasx_file_server_service import AasxFileServerService
 from basyx import ObjectStore
