--- api/test/examples/submodels.py
@@ -1,74 +1,0 @@
-test_submodel = {
-        "id": "urn:x-test:submodel1",
-        "submodelElements": [
-            {
-                "idShort": "some_property",
-                "valueType": "xs:int",
-                "value": "1984",
-                "modelType": "Property"
-            },
-            {
-                "idShort": "some_blob",
-                "value": "3q2+7w==",
-                "contentType": "application/octet-stream",
-                "modelType": "Blob"
-            },
-            {
-                "idShort": "ExampleSubmodelList",
-                "typeValueListElement": "SubmodelElementList",
-                "value": [
-                    {
-                        "idShort": "list_1",
-                        "value": "3q2+7w==",
-                        "contentType": "application/octet-stream",
-                        "modelType": "Blob"
-                    },
-                    {
-                        "idShort": "list_2",
-                        "value": "3q2+7w==",
-                        "contentType": "application/octet-stream",
-                        "modelType": "Blob"
-                    }
-                ],
-                "modelType": "SubmodelElementList"
-            }
-        ],
-        "modelType": "Submodel"
-    }
-test_submodel_modified = {
-    "id": "urn:x-test:submodel1",
-    "submodelElements": [
-        {
-            "idShort": "some_property",
-            "valueType": "xs:int",
-            "value": "8419",
-            "modelType": "Property"
-        },
-        {
-            "idShort": "some_blob",
-            "value": "3q2+7w==",
-            "contentType": "application/octet-stream",
-            "modelType": "Blob"
-        },
-        {
-            "idShort": "ExampleSubmodelList",
-            "typeValueListElement": "SubmodelElementList",
-            "value": [
-                {
-                    "idShort": "list_1",
-                    "value": "481563",
-                    "contentType": "application/octet-stream",
-                    "modelType": "Blob"
-                },
-                {
-                    "idShort": "list_2",
-                    "value": "&/6453=(",
-                    "contentType": "application/octet-stream",
-                    "modelType": "Blob"
-                }
-            ],
-            "modelType": "SubmodelElementList"
-        }
-    ],
-    "modelType": "Submodel"
-}
